week12/blog_api/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
